+++ src/Concordium/Client/Runner.hs
@@ -1,7 +1,6 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE NumericUnderscores #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE QuasiQuotes #-}
@@ -2857,40 +2856,14 @@ processConsensusCmd action _baseCfgDir verbose backend =
                 case res of
                     Left err -> logFatal ["Error getting chain parameters: " <> err]
                     Right Queries.EChainParametersAndKeys{..} -> runPrinter $ printChainParameters ecpParams
-        ConsensusChainUpdate rawUpdateFile keysFiles maybePLTParam intOpts -> do
+        ConsensusChainUpdate rawUpdateFile keysFiles intOpts -> do
             let
                 loadJSON :: (FromJSON a) => FilePath -> IO a
                 loadJSON fn =
                     AE.eitherDecodeFileStrict fn >>= \case
                         Left err -> logFatal [fn ++ ": " ++ err]
                         Right r -> return r
-            rawUpdate@Updates.RawUpdateInstruction{ruiPayload, ruiSeqNumber, ruiEffectiveTime, ruiTimeout} <- loadJSON rawUpdateFile
-            -- In case of a `CreatePLT` tx, load the init param from a separate JSON file and replace it for the `dummyValue` in the payload.
-            updatedRawUpdate <- case ruiPayload of
-                Updates.CreatePLTUpdatePayload (Types.CreatePLT{_cpltTokenSymbol, _cpltTokenModule, _cpltGovernanceAccount, _cpltDecimals, _cpltInitializationParameters}) ->
-                    case maybePLTParam of
-                        Just param -> do
-                            -- Load the new initialization parameters from the file.
-                            newInitializationParams <- loadJSON param
-                            let newCreatePLTPaylaod =
-                                    Types.CreatePLT
-                                        { _cpltTokenSymbol = _cpltTokenSymbol,
-                                          _cpltInitializationParameters = newInitializationParams,
-                                          _cpltTokenModule = _cpltTokenModule,
-                                          _cpltGovernanceAccount = _cpltGovernanceAccount,
-                                          _cpltDecimals = _cpltDecimals
-                                        }
-                            -- Replace the `dead` value.
-                            return
-                                rawUpdate
-                                    { Updates.ruiSeqNumber = ruiSeqNumber,
-                                      Updates.ruiEffectiveTime = ruiEffectiveTime,
-                                      Updates.ruiTimeout = ruiTimeout,
-                                      Updates.ruiPayload = Updates.CreatePLTUpdatePayload newCreatePLTPaylaod
-                                    }
-                        Nothing ->
-                            logFatal [printf "Creating a PLT token requires the `parameter-json` option."]
-                _ -> return rawUpdate
+            rawUpdate@Updates.RawUpdateInstruction{..} <- loadJSON rawUpdateFile
 
             Queries.EChainParametersAndKeys{..} <- withClient backend $ do
                 bcpRes <- getBlockChainParameters Best
@@ -2927,7 +2900,7 @@ processConsensusCmd action _baseCfgDir verbose backend =
             when (length keys < fromIntegral th) $
                 logFatal [printf "Not enough keys provided for signing this operation, got %u, need %u" (length keys) (fromIntegral th :: Int)]
             keyMap <- Map.fromList <$> mapM keyLU keys
-            let ui = Updates.makeUpdateInstruction updatedRawUpdate keyMap
+            let ui = Updates.makeUpdateInstruction rawUpdate keyMap
             when verbose $ logInfo ["Generated update instruction:", show ui]
             now <- getCurrentTimeUnix
             let expiryOK = ruiTimeout > now